routing set up GET to retrieve stitches from db. in progress routing to POST prototype images

+++ src/main/java/com/stitchingRetriever/stitchBuilderAPI/stitchBuilderAPI/model/FlossRepository.java
@@ -2,7 +2,7 @@ package com.stitchingRetriever.stitchBuilderAPI.stitchBuilderAPI.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FlossRepository extends JpaRepository<Floss, Long> {
+public interface FlossRepository extends JpaRepository<Floss, Integer> {
 
 
 }
